BUGFIX TODO 1. Subscribe to the same topic in one procedure will be wrong 2. When a subscriber exit, there may be one message left in the shared memory space.
1. Due to the implemention of normal ROS subscriber, it will be something
wrong when subscribe to the same topic in one procedure, for the
publisher will think there is only one subscriber listen to the topic,
but actually there are two. Maybe we can open an issue on github ros_comm.

2. When a subscriber exit, the publisher may not know the
decreace of the number of subscribers, so it still publish
more message than exactly what is. Then the extra message
will be left in the shared memory space. If the space is not
that big or subscribers exit many times, it will throw bad_alloc exception.